+++ api/src/anon_xfr/circuits.rs
@@ -11,7 +11,7 @@ use zei_algebra::{
     traits::{Group, Scalar},
     One, Zero,
 };
-use zei_crypto::basics::rescue::RescueInstance;
+use zei_crypto::basic::rescue::RescueInstance;
 use zei_crypto::{
     field_simulation::{SimFr, BIT_PER_LIMB, NUM_OF_LIMBS},
     pc_eq_rescue_split_verifier_zk_part::{NonZKState, ZKPartProof},
@@ -843,8 +843,8 @@ pub(crate) mod tests {
     use rand_core::SeedableRng;
     use ruc::*;
     use zei_algebra::{bls12_381::BLSScalar, traits::Scalar};
-    use zei_crypto::basics::rescue::RescueInstance;
-    use zei_crypto::basics::ristretto_pedersen_comm::RistrettoPedersenCommitment;
+    use zei_crypto::basic::rescue::RescueInstance;
+    use zei_crypto::basic::ristretto_pedersen_comm::RistrettoPedersenCommitment;
     use zei_crypto::pc_eq_rescue_split_verifier_zk_part::prove_pc_eq_rescue_external;
     use zei_plonk::plonk::constraint_system::{ecc::Point, TurboConstraintSystem};
