+++ api/src/anon_xfr/keys.rs
@@ -5,7 +5,7 @@ use zei_algebra::{
     jubjub::{JubjubPoint, JubjubScalar, JUBJUB_SCALAR_LEN},
     prelude::*,
 };
-use zei_crypto::basics::schnorr::{KeyPair, PublicKey, Signature};
+use zei_crypto::basic::schnorr::{KeyPair, PublicKey, Signature};
 
 const AXFR_SECRET_KEY_LENGTH: usize = JUBJUB_SCALAR_LEN;
 const AXFR_PUBLIC_KEY_LENGTH: usize = JubjubPoint::COMPRESSED_LEN;
