+++ crypto/src/pc_eq_rescue_split_verifier_zk_part.rs
@@ -1,5 +1,5 @@
-use crate::basics::rescue::RescueInstance;
-use crate::basics::ristretto_pedersen_comm::RistrettoPedersenCommitment;
+use crate::basic::rescue::RescueInstance;
+use crate::basic::ristretto_pedersen_comm::RistrettoPedersenCommitment;
 use crate::field_simulation::{SimFr, BIT_PER_LIMB, NUM_OF_LIMBS};
 use merlin::Transcript;
 use num_bigint::BigUint;
@@ -217,8 +217,8 @@ pub fn verify_pc_eq_rescue_external(
 
 #[cfg(test)]
 mod test {
-    use crate::basics::rescue::RescueInstance;
-    use crate::basics::ristretto_pedersen_comm::RistrettoPedersenCommitment;
+    use crate::basic::rescue::RescueInstance;
+    use crate::basic::ristretto_pedersen_comm::RistrettoPedersenCommitment;
     use crate::pc_eq_rescue_split_verifier_zk_part::{
         prove_pc_eq_rescue_external, verify_pc_eq_rescue_external,
     };
